+++ dev/examples/autoSiblingToChild.tsx
@@ -58,8 +58,8 @@ const Big = styled(motion.div)`
 
 const Child = () => {
     return (
-        <Big auto autoId="big" purple>
-            <Small auto autoId="small" purple />
+        <Big magic magicId="big" purple>
+            <Small magic magicId="small" purple />
         </Big>
     )
 }
@@ -67,8 +67,8 @@ const Child = () => {
 const Sibling = () => {
     return (
         <>
-            <Big auto autoId="big" />
-            <Small auto autoId="small" />
+            <Big magic magicId="big" />
+            <Small magic magicId="small" />
         </>
     )
 }